requirements.txt
PyNaCl
discord.py
aiohttp
torch
detoxify
ftfy
tqdm
async-timeout
